Terrain Generation done

+++ src/com/dylanscode/math/GameMath.java
@@ -2,10 +2,13 @@ package com.dylanscode.math;
 
 public class GameMath
 {
-	public static double distance(double x1,double y1, double x2, double y2) {
-		return Math.sqrt(Math.pow(x1-x2, 2)+Math.pow(y1-y2, 2));
+	public static double distance(double x1, double y1, double x2, double y2)
+	{
+		return Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2));
 	}
-	public static double norm(double o, double b, double e) {
-		return (o - b)/(e-b);
+
+	public static float norm(float o, float b, float e)
+	{
+		return (o - b) / (e - b);
 	}
 }
